commit: Sort the facets in descending order using arrrange function.   Author: Tammy Le   Date:   Sun March 3 7:50 PM
I added arrange(desc(Total_Injuries)) to sort the junction types by the total number of injuries in descending order. Then, I used this ordering to set the levels of the JUNCTIONTYPE factor, which facet_wrap uses to arrange the panels.

also, I added scales = "free_y" to facet_wrap to allow each facet to have its own Y-axis scale. i am playing around with this to make it easier to compare within each facet but be aware that this can sometimes make it harder to compare across facets.